統合してエラー2個まで減らした

+++ Shooters!/Vec2.cpp
@@ -17,22 +17,22 @@ Vec2 operator-(const Vec2 & v1, const Vec2 & v2){
 
 Vec2 operator*(const double d, const Vec2 & v1){
 	Vec2 v;
-	v.x = d * v1.x;
-	v.y = d * v1.y;
+	v.x = (float)d * v1.x;
+	v.y = (float)d * v1.y;
 	return v;
 }
 
 Vec2 operator*(const Vec2 & v1, const double d){
 	Vec2 v;
-	v.x = d * v1.x;
-	v.y = d * v1.y;
+	v.x = (float)d * v1.x;
+	v.y = (float)d * v1.y;
 	return v;
 }
 
 Vec2 operator/(const Vec2 & v1, const double d) {
 	Vec2 v;
-	v.x = v1.x / d;
-	v.y = v1.y / d;
+	v.x = v1.x / (float)d;
+	v.y = v1.y / (float)d;
 	return v;
 }
